refactor(frontend): split run page into ReviewPanel/run-widgets
