+++ frontend/src/app/run/[sessionId]/page.tsx
@@ -25,13 +25,9 @@ import {
   RUN_STALL_TIMEOUT_MS,
   RUNSERVER_MARKER_AUTHOR,
   PIPELINE_STATE_KEYS,
-  FIELD_LABELS,
-  FIELD_COLORS,
-  HIDDEN_FIELDS,
-  WIDGET_LAYOUTS,
-  DEFAULT_LAYOUT,
 } from "./run-config";
-import { widgetAccent, extractItems, parseRawGtrends } from "./run-helpers";
+import { PipelineWidget } from "./run-widgets";
+import { ReviewPanel } from "./ReviewPanel";
 
 type Status = "running" | "completed" | "error" | "paused" | "stalled";
 
@@ -39,561 +35,6 @@ interface PauseContext {
   functionCallId: string;
   functionName: string;
   eventId: string;
-}
-
-/** Render a single field with label + value. */
-function FieldCell({ fieldKey, value }: { fieldKey: string; value: unknown }) {
-  const label = FIELD_LABELS[fieldKey] || fieldKey.replace(/_/g, " ");
-  const labelColor = FIELD_COLORS[fieldKey] || "text-muted-foreground";
-  return (
-    <div>
-      <dt className={`text-[10px] font-bold uppercase tracking-wider ${labelColor}`}>
-        {label}
-      </dt>
-      <dd className="text-xs leading-snug break-words whitespace-pre-wrap text-foreground/85 mt-0.5">
-        {typeof value === "string" || typeof value === "number"
-          ? String(value)
-          : JSON.stringify(value, null, 2)}
-      </dd>
-    </div>
-  );
-}
-
-/** Render a single item card with side-by-side panels layout. */
-function ItemCard({
-  item,
-  index,
-  widgetKey,
-}: {
-  item: Record<string, unknown>;
-  index: number;
-  widgetKey: string;
-}): React.ReactNode {
-  const title =
-    (item.concept_name as string) ||
-    (item.headline as string) ||
-    `Item ${index + 1}`;
-
-  const layout = WIDGET_LAYOUTS[widgetKey] || DEFAULT_LAYOUT;
-  const pairs: [string, string][] = layout.pairs;
-  const fullWidth = layout.fullWidth;
-
-  // Collect fields used in structured layout
-  const structuredFields = new Set<string>();
-  pairs.forEach(([a, b]) => { structuredFields.add(a); structuredFields.add(b); });
-  if (fullWidth) structuredFields.add(fullWidth);
-  structuredFields.add("concept_name");
-  structuredFields.add("headline");
-  HIDDEN_FIELDS.forEach((f) => structuredFields.add(f));
-
-  // Remaining fields not covered by structured layout
-  const remainingEntries = Object.entries(item).filter(
-    ([k, v]) => v !== null && v !== undefined && v !== "" && !structuredFields.has(k)
-  );
-
-  return (
-    <div className="glass rounded-xl px-4 py-3 space-y-1.5">
-      {/* Title bar */}
-      <div className="flex items-center gap-2 border-b border-border pb-2">
-        <Badge
-          variant="secondary"
-          className="text-[10px] px-1.5 py-0 bg-primary/10 text-primary border-0 font-bold"
-        >
-          {index + 1}
-        </Badge>
-        <span className="text-sm font-bold leading-tight text-foreground">
-          {title}
-        </span>
-      </div>
-
-      {/* Side-by-side panels */}
-      {pairs.map(([leftKey, rightKey]): React.ReactNode => {
-        const leftVal = item[leftKey];
-        const rightVal = item[rightKey];
-        if (!leftVal && !rightVal) return null;
-        return (
-          <div key={`${leftKey}-${rightKey}`} className="grid grid-cols-2 gap-2">
-            <div className="rounded-md bg-muted/50 px-2.5 py-2">
-              {leftVal ? (
-                <FieldCell fieldKey={leftKey} value={leftVal} />
-              ) : (
-                <span className="text-xs text-muted-foreground italic">--</span>
-              )}
-            </div>
-            <div className="rounded-md bg-muted/50 px-2.5 py-2">
-              {rightVal ? (
-                <FieldCell fieldKey={rightKey} value={rightVal} />
-              ) : (
-                <span className="text-xs text-muted-foreground italic">--</span>
-              )}
-            </div>
-          </div>
-        );
-      })}
-
-      {/* Full-width panel */}
-      {fullWidth && !!item[fullWidth] && (
-        <div className="rounded-md bg-cyan-50/60 border border-cyan-200/40 px-2.5 py-2">
-          <FieldCell fieldKey={fullWidth} value={item[fullWidth]} />
-        </div>
-      )}
-
-      {/* Remaining fields not in the structured layout */}
-      {remainingEntries.length > 0 && (
-        <div className="space-y-1.5 pt-1 border-t border-border">
-          {remainingEntries.map(([key, value]) => (
-            <FieldCell key={key} fieldKey={key} value={value} />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function PipelineWidget({
-  label,
-  stateKey,
-  data,
-}: {
-  label: string;
-  stateKey: string;
-  data: unknown;
-}) {
-  const [open, setOpen] = useState(false);
-  const items = extractItems(data);
-  const itemCount = items ? items.length : 0;
-  const accent = widgetAccent(label);
-
-  return (
-    <>
-      <div
-        className="glass rounded-xl cursor-pointer transition-all duration-200
-                   hover:shadow-md hover:shadow-black/5
-                   animate-fadeInUpSmooth"
-        onClick={() => setOpen(true)}
-      >
-        <div className="py-2.5 px-4 flex items-center justify-between">
-          <span className="flex items-center gap-2 text-sm font-bold">
-            <span className={`inline-block h-2 w-2 rounded-full ${accent.dot}`} />
-            <span className={accent.text}>{label}</span>
-            {itemCount > 0 && (
-              <Badge
-                variant="secondary"
-                className={`text-[10px] px-1.5 py-0 border font-bold ${accent.badge}`}
-              >
-                {itemCount}
-              </Badge>
-            )}
-          </span>
-          <span className="text-[10px] text-muted-foreground">
-            click to view
-          </span>
-        </div>
-      </div>
-
-      {open && (
-        <div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-black/30 backdrop-blur-sm animate-in fade-in duration-150"
-          onClick={() => setOpen(false)}
-        >
-          <div
-            className="relative mx-4 w-full max-w-3xl max-h-[85vh] overflow-hidden flex flex-col
-                       glass-strong rounded-2xl shadow-2xl shadow-black/10
-                       animate-in zoom-in-95 duration-200"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="flex items-center justify-between border-b border-border px-5 py-4">
-              <h2 className="text-lg font-bold flex items-center gap-2">
-                <span className={`inline-block h-2.5 w-2.5 rounded-full ${accent.dot}`} />
-                <span className={accent.text}>{label}</span>
-                {itemCount > 0 && (
-                  <Badge
-                    variant="secondary"
-                    className={`border font-bold ${accent.badge}`}
-                  >
-                    {itemCount} items
-                  </Badge>
-                )}
-              </h2>
-              <button
-                onClick={() => setOpen(false)}
-                className="flex h-7 w-7 items-center justify-center rounded-lg text-muted-foreground hover:text-foreground hover:bg-black/5 transition-colors"
-                aria-label="Close"
-              >
-                &times;
-              </button>
-            </div>
-            <div className="flex-1 overflow-auto p-5">
-              {items ? (
-                <div className="space-y-3">
-                  {items.map((item, i) => (
-                    <ItemCard key={i} item={item} index={i} widgetKey={stateKey} />
-                  ))}
-                </div>
-              ) : (
-                <pre className="whitespace-pre-wrap break-words rounded-lg bg-muted/50 p-4 text-sm font-mono leading-relaxed text-foreground/80">
-                  {typeof data === "string"
-                    ? data
-                    : JSON.stringify(data, null, 2)}
-                </pre>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-    </>
-  );
-}
-
-/* ── Review panel components for interactive mode ── */
-
-function ReviewResearch({
-  state,
-  onResume,
-}: {
-  state: Record<string, unknown>;
-  onResume: (response: Record<string, unknown>) => void;
-}) {
-  const [feedback, setFeedback] = useState("");
-  const [editMode, setEditMode] = useState(false);
-  const report = state.combined_final_cited_report as string | undefined;
-  const [editedReport, setEditedReport] = useState(report ?? "");
-
-  return (
-    <div className="glass rounded-2xl p-6 space-y-4 animate-fadeInUp">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <span className="inline-block h-2.5 w-2.5 rounded-full bg-amber-500" />
-          <h2 className="text-lg font-semibold">Review Research Report</h2>
-        </div>
-        {report && (
-          <button
-            onClick={() => setEditMode((v) => !v)}
-            className="text-xs font-medium text-primary hover:text-primary/80 transition-colors"
-          >
-            {editMode ? "Preview" : "Edit"}
-          </button>
-        )}
-      </div>
-      <p className="text-sm text-muted-foreground">
-        Review the research findings below. Approve to continue to ad copy generation,
-        or provide feedback for revisions.
-      </p>
-      {report && !editMode && (
-        <div className="max-h-[28rem] overflow-y-auto rounded-lg bg-muted/30 p-5 border border-border prose prose-sm prose-neutral max-w-none
-          prose-headings:text-foreground prose-headings:font-bold
-          prose-h1:text-lg prose-h2:text-base prose-h3:text-sm
-          prose-p:text-foreground/85 prose-p:leading-relaxed
-          prose-a:text-primary prose-a:no-underline hover:prose-a:underline
-          prose-strong:text-foreground prose-strong:font-semibold
-          prose-li:text-foreground/85
-          prose-code:text-xs prose-code:bg-muted prose-code:rounded prose-code:px-1
-          prose-blockquote:border-l-primary/30 prose-blockquote:text-muted-foreground">
-          <ReactMarkdown remarkPlugins={[remarkGfm]}>{report}</ReactMarkdown>
-        </div>
-      )}
-      {report && editMode && (
-        <Textarea
-          value={editedReport}
-          onChange={(e) => setEditedReport(e.target.value)}
-          rows={16}
-          className="bg-background border-border font-mono text-xs leading-relaxed max-h-[28rem]"
-        />
-      )}
-      <Textarea
-        placeholder="Optional feedback or revision requests..."
-        value={feedback}
-        onChange={(e) => setFeedback(e.target.value)}
-        rows={3}
-        className="bg-background border-border"
-      />
-      <div className="flex gap-3">
-        <Button onClick={() => onResume({ status: "approved", feedback, instruction: "User approved the research. Continue to the next step in the WORKFLOW." })}>
-          Approve &amp; Continue
-        </Button>
-        <Button
-          variant="outline"
-          onClick={() => onResume({ status: "revision_requested", feedback, instruction: "User requested changes to the research. Address their feedback, then continue the WORKFLOW." })}
-          disabled={!feedback}
-        >
-          Request Changes
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-/** Labeled field for review cards */
-function ReviewField({ label, value, color }: { label: string; value: string; color: string }) {
-  if (!value) return null;
-  return (
-    <div>
-      <dt className={`text-[10px] font-bold uppercase tracking-wider ${color}`}>{label}</dt>
-      <dd className="mt-0.5 text-sm leading-snug text-foreground/85">{value}</dd>
-    </div>
-  );
-}
-
-function ReviewAdCopies({
-  state,
-  onResume,
-}: {
-  state: Record<string, unknown>;
-  onResume: (response: Record<string, unknown>) => void;
-}) {
-  const [feedback, setFeedback] = useState("");
-  const adCopies = extractItems(state.ad_copy_critique);
-
-  return (
-    <div className="glass rounded-2xl p-6 space-y-4 animate-fadeInUp">
-      <div className="flex items-center gap-2">
-        <span className="inline-block h-2.5 w-2.5 rounded-full bg-amber-500" />
-        <h2 className="text-lg font-semibold">Review Ad Copies</h2>
-      </div>
-      <p className="text-sm text-muted-foreground">
-        Review the generated ad copies. Approve to continue to visual concept generation.
-      </p>
-      {adCopies && (
-        <div className="space-y-3 max-h-[28rem] overflow-y-auto">
-          {adCopies.map((copy, i) => (
-            <div key={i} className="rounded-xl border border-border/60 bg-background/50 p-4 space-y-3">
-              {/* Title bar */}
-              <div className="flex items-center gap-2 pb-2 border-b border-border/40">
-                <Badge variant="secondary" className="text-[10px] px-1.5 py-0 bg-indigo-100 text-indigo-600 border-0 font-bold">
-                  {i + 1}
-                </Badge>
-                <span className="text-sm font-bold text-foreground">
-                  {String(copy.headline ?? `Ad Copy ${i + 1}`)}
-                </span>
-              </div>
-
-              {/* Two-column field grid */}
-              <div className="grid grid-cols-2 gap-3">
-                <div className="rounded-md bg-muted/40 px-3 py-2">
-                  <ReviewField label="Body Text" value={String(copy.body_text ?? "")} color="text-indigo-500" />
-                </div>
-                <div className="rounded-md bg-muted/40 px-3 py-2">
-                  <ReviewField label="Tone / Style" value={String(copy.tone_style ?? "")} color="text-violet-500" />
-                </div>
-                <div className="rounded-md bg-muted/40 px-3 py-2">
-                  <ReviewField label="Call to Action" value={String(copy.call_to_action ?? "")} color="text-amber-600" />
-                </div>
-                <div className="rounded-md bg-muted/40 px-3 py-2">
-                  <ReviewField label="Trend Connection" value={String(copy.trend_connection ?? "")} color="text-emerald-600" />
-                </div>
-              </div>
-
-              {/* Full-width fields */}
-              {!!copy.audience_appeal_rationale && (
-                <div className="rounded-md bg-pink-50/60 border border-pink-200/40 px-3 py-2">
-                  <ReviewField label="Audience Appeal" value={String(copy.audience_appeal_rationale)} color="text-pink-600" />
-                </div>
-              )}
-              {!!copy.social_caption && (
-                <div className="rounded-md bg-amber-50/60 border border-amber-200/40 px-3 py-2">
-                  <ReviewField label="Social Caption" value={String(copy.social_caption)} color="text-amber-500" />
-                </div>
-              )}
-              {!!copy.detailed_performance_rationale && (
-                <div className="rounded-md bg-orange-50/60 border border-orange-200/40 px-3 py-2">
-                  <ReviewField label="Performance Rationale" value={String(copy.detailed_performance_rationale)} color="text-orange-500" />
-                </div>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-      <Textarea
-        placeholder="Optional feedback..."
-        value={feedback}
-        onChange={(e) => setFeedback(e.target.value)}
-        rows={3}
-        className="bg-background border-border"
-      />
-      <div className="flex gap-3">
-        <Button onClick={() => onResume({ status: "approved", feedback, instruction: "User approved the ad copies. Continue to the next step in the WORKFLOW — generate visual concepts." })}>
-          Approve &amp; Continue
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ReviewVisualConcepts({
-  state,
-  onResume,
-}: {
-  state: Record<string, unknown>;
-  onResume: (response: Record<string, unknown>) => void;
-}) {
-  const [feedback, setFeedback] = useState("");
-  const concepts = extractItems(state.final_visual_concepts);
-
-  return (
-    <div className="glass rounded-2xl p-6 space-y-4 animate-fadeInUp">
-      <div className="flex items-center gap-2">
-        <span className="inline-block h-2.5 w-2.5 rounded-full bg-amber-500" />
-        <h2 className="text-lg font-semibold">Review Visual Concepts</h2>
-      </div>
-      <p className="text-sm text-muted-foreground">
-        Review the visual concepts and image prompts. Approve to generate images.
-      </p>
-      {concepts && (
-        <div className="space-y-3 max-h-96 overflow-y-auto">
-          {concepts.map((concept, i) => (
-            <div key={i} className="rounded-lg border p-4 space-y-2">
-              <div className="font-medium">{String(concept.concept_name ?? `Concept ${i + 1}`)}</div>
-              <div className="text-sm text-muted-foreground">{String(concept.concept_summary ?? "")}</div>
-              <div className="text-xs text-cyan-600 font-mono bg-muted/30 rounded p-2">
-                {String(concept.image_generation_prompt ?? "")}
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-      <Textarea
-        placeholder="Optional feedback..."
-        value={feedback}
-        onChange={(e) => setFeedback(e.target.value)}
-        rows={3}
-        className="bg-background border-border"
-      />
-      <div className="flex gap-3">
-        <Button onClick={() => onResume({ status: "approved", feedback, instruction: "User approved the visual concepts. Continue to the next step in the WORKFLOW — generate images." })}>
-          Approve &amp; Generate Images
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ReviewTrends({
-  state,
-  onResume,
-}: {
-  state: Record<string, unknown>;
-  onResume: (response: Record<string, unknown>) => void;
-}) {
-  const candidates = useMemo(
-    () => parseRawGtrends(state.raw_gtrends),
-    [state.raw_gtrends]
-  );
-  const [selected, setSelected] = useState<Set<string>>(new Set());
-  const [instruction, setInstruction] = useState("");
-
-  const toggle = (term: string) => {
-    setSelected((prev) => {
-      const next = new Set(prev);
-      if (next.has(term)) next.delete(term);
-      else next.add(term);
-      return next;
-    });
-  };
-
-  return (
-    <div className="glass rounded-2xl p-6 space-y-4 animate-fadeInUp">
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <span className="inline-block h-2.5 w-2.5 rounded-full bg-amber-500" />
-          <h2 className="text-lg font-semibold">Pick Your Trends</h2>
-        </div>
-        {candidates.length > 0 && (
-          <Badge
-            variant="secondary"
-            className="text-[10px] px-1.5 py-0 bg-primary/10 text-primary border-0 font-bold"
-          >
-            {selected.size} / {candidates.length} selected
-          </Badge>
-        )}
-      </div>
-      <p className="text-sm text-muted-foreground">
-        Select the trends you want to keep. The agent will skip its automatic
-        pick and research the trends you choose.
-      </p>
-      {candidates.length === 0 ? (
-        <p className="text-sm text-muted-foreground italic">
-          No candidate trends available yet.
-        </p>
-      ) : (
-        <div className="grid gap-2 sm:grid-cols-2 max-h-[28rem] overflow-y-auto">
-          {candidates.map((term) => {
-            const isSelected = selected.has(term);
-            return (
-              <button
-                key={term}
-                type="button"
-                onClick={() => toggle(term)}
-                aria-pressed={isSelected}
-                className={`glass rounded-xl px-4 py-3 text-left text-sm transition-all duration-200 hover:shadow-md hover:shadow-black/5 ${
-                  isSelected
-                    ? "ring-2 ring-primary bg-primary/10 text-primary font-semibold"
-                    : "text-foreground/85"
-                }`}
-              >
-                <span className="flex items-center gap-2">
-                  <span
-                    className={`inline-flex h-4 w-4 shrink-0 items-center justify-center rounded border text-[10px] font-bold ${
-                      isSelected
-                        ? "border-primary bg-primary text-primary-foreground"
-                        : "border-border text-transparent"
-                    }`}
-                  >
-                    &#10003;
-                  </span>
-                  <span className="leading-snug break-words">{term}</span>
-                </span>
-              </button>
-            );
-          })}
-        </div>
-      )}
-      <Textarea
-        placeholder="Optional note for the agent (e.g. focus, angle)..."
-        value={instruction}
-        onChange={(e) => setInstruction(e.target.value)}
-        rows={2}
-        className="bg-background border-border"
-      />
-      <div className="flex gap-3">
-        <Button
-          disabled={selected.size === 0}
-          onClick={() =>
-            onResume({
-              status: "selected",
-              selected_trends: [...selected],
-              instruction,
-            })
-          }
-        >
-          Confirm Selection
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ReviewPanel({
-  functionName,
-  sessionState,
-  onResume,
-}: {
-  functionName: string;
-  sessionState: Record<string, unknown>;
-  onResume: (response: Record<string, unknown>) => void;
-}) {
-  if (functionName === "review_research") {
-    return <ReviewResearch state={sessionState} onResume={onResume} />;
-  }
-  if (functionName === "review_ad_copies") {
-    return <ReviewAdCopies state={sessionState} onResume={onResume} />;
-  }
-  if (functionName === "review_visual_concepts") {
-    return <ReviewVisualConcepts state={sessionState} onResume={onResume} />;
-  }
-  if (functionName === "review_trends") {
-    return <ReviewTrends state={sessionState} onResume={onResume} />;
-  }
-  return null;
 }
 
 export default function RunPage({
